+++ http/server.ts
@@ -9,7 +9,6 @@ import {
   readRequest,
   writeResponse,
 } from "./_io.ts";
-
 export class ServerRequest {
   url!: string;
   method!: string;
@@ -55,7 +54,7 @@ export class ServerRequest {
   /**
    * Body of the request.  The easiest way to consume the body is:
    *
-   *     const buf: Uint8Array = await Deno.readAll(req.body);
+   *     const buf: Uint8Array = await readAll(req.body);
    */
   get body(): Deno.Reader {
     if (!this.#body) {
@@ -160,7 +159,7 @@ export class Server implements AsyncIterable<ServerRequest> {
               status: 400,
               body: new TextEncoder().encode(`${error.message}\r\n\r\n`),
             });
-          } catch (error) {
+          } catch {
             // The connection is broken.
           }
         }
@@ -187,7 +186,7 @@ export class Server implements AsyncIterable<ServerRequest> {
       try {
         // Consume unread body and trailers if receiver didn't consume those data
         await request.finalize();
-      } catch (error) {
+      } catch {
         // Invalid data was received or the connection was closed.
         break;
       }
@@ -196,7 +195,7 @@ export class Server implements AsyncIterable<ServerRequest> {
     this.untrackConnection(conn);
     try {
       conn.close();
-    } catch (e) {
+    } catch {
       // might have been already closed
     }
   }