+++ http/server_test.ts
@@ -25,6 +25,8 @@ import { BufReader, BufWriter } from "../io/bufio.ts";
 import { delay } from "../async/delay.ts";
 import { mockConn } from "./_mock_conn.ts";
 import { dirname, fromFileUrl, join, resolve } from "../path/mod.ts";
+import { Buffer } from "../io/buffer.ts";
+import { readAll, writeAll } from "../io/util.ts";
 
 const moduleDir = dirname(fromFileUrl(import.meta.url));
 const testdataDir = resolve(moduleDir, "testdata");
@@ -51,7 +53,7 @@ const responseTests: ResponseTest[] = [
   {
     response: {
       status: 200,
-      body: new Deno.Buffer(new TextEncoder().encode("abcdef")),
+      body: new Buffer(new TextEncoder().encode("abcdef")),
     },
 
     raw: "HTTP/1.1 200 OK\r\n" +
@@ -62,7 +64,7 @@ const responseTests: ResponseTest[] = [
 
 Deno.test("responseWrite", async function (): Promise<void> {
   for (const testCase of responseTests) {
-    const buf = new Deno.Buffer();
+    const buf = new Buffer();
     const bufw = new BufWriter(buf);
     const request = new ServerRequest();
     request.w = bufw;
@@ -81,7 +83,7 @@ Deno.test("requestContentLength", function (): void {
     const req = new ServerRequest();
     req.headers = new Headers();
     req.headers.set("content-length", "5");
-    const buf = new Deno.Buffer(new TextEncoder().encode("Hello"));
+    const buf = new Buffer(new TextEncoder().encode("Hello"));
     req.r = new BufReader(buf);
     assertEquals(req.contentLength, 5);
   }
@@ -102,7 +104,7 @@ Deno.test("requestContentLength", function (): void {
       chunkOffset += chunkSize;
     }
     chunksData += "0\r\n\r\n";
-    const buf = new Deno.Buffer(new TextEncoder().encode(chunksData));
+    const buf = new Buffer(new TextEncoder().encode(chunksData));
     req.r = new BufReader(buf);
     assertEquals(req.contentLength, null);
   }
@@ -132,9 +134,9 @@ Deno.test("requestBodyWithContentLength", async function (): Promise<void> {
     const req = new ServerRequest();
     req.headers = new Headers();
     req.headers.set("content-length", "5");
-    const buf = new Deno.Buffer(new TextEncoder().encode("Hello"));
+    const buf = new Buffer(new TextEncoder().encode("Hello"));
     req.r = new BufReader(buf);
-    const body = new TextDecoder().decode(await Deno.readAll(req.body));
+    const body = new TextDecoder().decode(await readAll(req.body));
     assertEquals(body, "Hello");
   }
 
@@ -144,9 +146,9 @@ Deno.test("requestBodyWithContentLength", async function (): Promise<void> {
     const req = new ServerRequest();
     req.headers = new Headers();
     req.headers.set("Content-Length", "5000");
-    const buf = new Deno.Buffer(new TextEncoder().encode(longText));
+    const buf = new Buffer(new TextEncoder().encode(longText));
     req.r = new BufReader(buf);
-    const body = new TextDecoder().decode(await Deno.readAll(req.body));
+    const body = new TextDecoder().decode(await readAll(req.body));
     assertEquals(body, longText);
   }
   // Handler ignored to consume body
@@ -158,9 +160,9 @@ Deno.test(
     const req = new ServerRequest();
     req.headers = new Headers();
     req.headers.set("content-length", "" + text.length);
-    const tr = totalReader(new Deno.Buffer(new TextEncoder().encode(text)));
+    const tr = totalReader(new Buffer(new TextEncoder().encode(text)));
     req.r = new BufReader(tr);
-    req.w = new BufWriter(new Deno.Buffer());
+    req.w = new BufWriter(new Buffer());
     await req.respond({ status: 200, body: "ok" });
     assertEquals(tr.total, 0);
     await req.finalize();
@@ -187,9 +189,9 @@ Deno.test(
     req.headers.set("transfer-encoding", "chunked");
     req.headers.set("trailer", "deno,node");
     const body = new TextEncoder().encode(text);
-    const tr = totalReader(new Deno.Buffer(body));
+    const tr = totalReader(new Buffer(body));
     req.r = new BufReader(tr);
-    req.w = new BufWriter(new Deno.Buffer());
+    req.w = new BufWriter(new Buffer());
     await req.respond({ status: 200, body: "ok" });
     assertEquals(tr.total, 0);
     assertEquals(req.headers.has("trailer"), true);
@@ -219,9 +221,9 @@ Deno.test("requestBodyWithTransferEncoding", async function (): Promise<void> {
       chunkOffset += chunkSize;
     }
     chunksData += "0\r\n\r\n";
-    const buf = new Deno.Buffer(new TextEncoder().encode(chunksData));
+    const buf = new Buffer(new TextEncoder().encode(chunksData));
     req.r = new BufReader(buf);
-    const body = new TextDecoder().decode(await Deno.readAll(req.body));
+    const body = new TextDecoder().decode(await readAll(req.body));
     assertEquals(body, shortText);
   }
 
@@ -242,9 +244,9 @@ Deno.test("requestBodyWithTransferEncoding", async function (): Promise<void> {
       chunkOffset += chunkSize;
     }
     chunksData += "0\r\n\r\n";
-    const buf = new Deno.Buffer(new TextEncoder().encode(chunksData));
+    const buf = new Buffer(new TextEncoder().encode(chunksData));
     req.r = new BufReader(buf);
-    const body = new TextDecoder().decode(await Deno.readAll(req.body));
+    const body = new TextDecoder().decode(await readAll(req.body));
     assertEquals(body, longText);
   }
 });
@@ -257,7 +259,7 @@ Deno.test("requestBodyReaderWithContentLength", async function (): Promise<
     const req = new ServerRequest();
     req.headers = new Headers();
     req.headers.set("content-length", "" + shortText.length);
-    const buf = new Deno.Buffer(new TextEncoder().encode(shortText));
+    const buf = new Buffer(new TextEncoder().encode(shortText));
     req.r = new BufReader(buf);
     const readBuf = new Uint8Array(6);
     let offset = 0;
@@ -278,7 +280,7 @@ Deno.test("requestBodyReaderWithContentLength", async function (): Promise<
     const req = new ServerRequest();
     req.headers = new Headers();
     req.headers.set("Content-Length", "5000");
-    const buf = new Deno.Buffer(new TextEncoder().encode(longText));
+    const buf = new Buffer(new TextEncoder().encode(longText));
     req.r = new BufReader(buf);
     const readBuf = new Uint8Array(1000);
     let offset = 0;
@@ -313,7 +315,7 @@ Deno.test("requestBodyReaderWithTransferEncoding", async function (): Promise<
       chunkOffset += chunkSize;
     }
     chunksData += "0\r\n\r\n";
-    const buf = new Deno.Buffer(new TextEncoder().encode(chunksData));
+    const buf = new Buffer(new TextEncoder().encode(chunksData));
     req.r = new BufReader(buf);
     const readBuf = new Uint8Array(6);
     let offset = 0;
@@ -345,7 +347,7 @@ Deno.test("requestBodyReaderWithTransferEncoding", async function (): Promise<
       chunkOffset += chunkSize;
     }
     chunksData += "0\r\n\r\n";
-    const buf = new Deno.Buffer(new TextEncoder().encode(chunksData));
+    const buf = new Buffer(new TextEncoder().encode(chunksData));
     req.r = new BufReader(buf);
     const readBuf = new Uint8Array(1000);
     let offset = 0;
@@ -445,7 +447,7 @@ Deno.test({
         port: 4503,
         certFile: join(testdataDir, "tls/RootCA.pem"),
       });
-      await Deno.writeAll(
+      await writeAll(
         conn,
         new TextEncoder().encode("GET / HTTP/1.0\r\n\r\n"),
       );
@@ -491,7 +493,7 @@ Deno.test({
     const server = serve(":8123");
     const p = iteratorReq(server);
     const conn = await Deno.connect({ hostname: "127.0.0.1", port: 8123 });
-    await Deno.writeAll(
+    await writeAll(
       conn,
       new TextEncoder().encode("GET /hello HTTP/1.1\r\n\r\n"),
     );
@@ -532,7 +534,7 @@ Deno.test({
       hostname: "127.0.0.1",
       port: 8124,
     });
-    await Deno.writeAll(
+    await writeAll(
       conn,
       new TextEncoder().encode("GET / HTTP/1.1\r\n\r\n"),
     );
@@ -550,13 +552,13 @@ Deno.test({
       hostname: "127.0.0.1",
       port: 8124,
     });
-    await Deno.writeAll(
+    await writeAll(
       conn,
       new TextEncoder().encode(
         "GET / HTTP/1.1\r\nmalformedHeader\r\n\r\n\r\n\r\n",
       ),
     );
-    const responseString = new TextDecoder().decode(await Deno.readAll(conn));
+    const responseString = new TextDecoder().decode(await readAll(conn));
     assertMatch(
       responseString,
       /^HTTP\/1\.1 400 Bad Request\r\ncontent-length: \d+\r\n\r\n.*\r\n\r\n$/ms,
@@ -583,14 +585,14 @@ Deno.test({
       hostname: "127.0.0.1",
       port: 8124,
     });
-    await Deno.writeAll(
+    await writeAll(
       conn,
       new TextEncoder().encode(
         "PUT / HTTP/1.1\r\nTransfer-Encoding: chunked\r\n\r\nzzzzzzz\r\nhello",
       ),
     );
     await conn.closeWrite();
-    const responseString = new TextDecoder().decode(await Deno.readAll(conn));
+    const responseString = new TextDecoder().decode(await readAll(conn));
     assertEquals(
       responseString,
       "HTTP/1.1 200 OK\r\ncontent-length: 13\r\n\r\nHello, world!",
@@ -616,14 +618,14 @@ Deno.test({
       hostname: "127.0.0.1",
       port: 8124,
     });
-    await Deno.writeAll(
+    await writeAll(
       conn,
       new TextEncoder().encode(
         "PUT / HTTP/1.1\r\nTransfer-Encoding: chunked\r\n\r\n5\r\nHello",
       ),
     );
     conn.closeWrite();
-    const responseString = new TextDecoder().decode(await Deno.readAll(conn));
+    const responseString = new TextDecoder().decode(await readAll(conn));
     assertEquals(
       responseString,
       "HTTP/1.1 200 OK\r\ncontent-length: 13\r\n\r\nHello, world!",
@@ -648,7 +650,7 @@ Deno.test({
       hostname: "127.0.0.1",
       port: 8124,
     });
-    await Deno.writeAll(
+    await writeAll(
       conn,
       new TextEncoder().encode([
         // A normal request is required:
@@ -719,7 +721,7 @@ Deno.test({
         certFile: join(testdataDir, "tls/RootCA.pem"),
       });
 
-      await Deno.writeAll(
+      await writeAll(
         conn,
         new TextEncoder().encode("GET / HTTP/1.0\r\n\r\n"),
       );