+++ io/streams_test.ts
@@ -2,11 +2,13 @@
 
 import { assert, assertEquals } from "../testing/asserts.ts";
 import {
-  readableStreamFromAsyncIterator,
+  readableStreamFromIterable,
+  readerFromIterable,
   readerFromStreamReader,
   writableStreamFromWriter,
   writerFromStreamWriter,
 } from "./streams.ts";
+import { Buffer } from "./buffer.ts";
 
 function repeat(c: string, bytes: number): Uint8Array {
   assertEquals(c.length, 1);
@@ -15,7 +17,28 @@ function repeat(c: string, bytes: number): Uint8Array {
   return ui8;
 }
 
-Deno.test("toWriterCheck", async function (): Promise<void> {
+Deno.test("[io] readerFromIterable()", async function () {
+  const reader = readerFromIterable((function* () {
+    const encoder = new TextEncoder();
+    for (const string of ["hello", "deno", "foo"]) {
+      yield encoder.encode(string);
+    }
+  })());
+
+  const readStrings = [];
+  const decoder = new TextDecoder();
+  const p = new Uint8Array(4);
+  while (true) {
+    const n = await reader.read(p);
+    if (n == null) {
+      break;
+    }
+    readStrings.push(decoder.decode(p.slice(0, n)));
+  }
+  assertEquals(readStrings, ["hell", "o", "deno", "foo"]);
+});
+
+Deno.test("[io] writerFromStreamWriter()", async function () {
   const written: string[] = [];
   const chunks: string[] = ["hello", "deno", "land"];
   const writableStream = new WritableStream({
@@ -37,7 +60,7 @@ Deno.test("toWriterCheck", async function (): Promise<void> {
   assertEquals(written, chunks);
 });
 
-Deno.test("toReaderCheck", async function (): Promise<void> {
+Deno.test("[io] readerFromStreamReader()", async function () {
   const chunks: string[] = ["hello", "deno", "land"];
   const expected = chunks.slice();
   const readChunks: Uint8Array[] = [];
@@ -73,10 +96,10 @@ Deno.test("toReaderCheck", async function (): Promise<void> {
   );
 });
 
-Deno.test("toReaderBigChunksCheck", async function (): Promise<void> {
+Deno.test("[io] readerFromStreamReader() big chunks", async function () {
   const bufSize = 1024;
   const chunkSize = 3 * bufSize;
-  const writer = new Deno.Buffer();
+  const writer = new Buffer();
 
   // A readable stream can enqueue chunks bigger than Copy bufSize
   // Reader returned by toReader should enqueue exceeding bytes
@@ -104,10 +127,10 @@ Deno.test("toReaderBigChunksCheck", async function (): Promise<void> {
   assertEquals(writer.length, expectedWritten);
 });
 
-Deno.test("toReaderBigIrregularChunksCheck", async function (): Promise<void> {
+Deno.test("[io] readerFromStreamReader() irregular chunks", async function () {
   const bufSize = 1024;
   const chunkSize = 3 * bufSize;
-  const writer = new Deno.Buffer();
+  const writer = new Buffer();
 
   // A readable stream can enqueue chunks bigger than Copy bufSize
   // Reader returned by toReader should enqueue exceeding bytes
@@ -138,7 +161,7 @@ Deno.test("toReaderBigIrregularChunksCheck", async function (): Promise<void> {
   assertEquals(expected, writer.bytes());
 });
 
-Deno.test("toWritableCheck", async function (): Promise<void> {
+Deno.test("[io] writableStreamFromWriter()", async function () {
   const written: string[] = [];
   const chunks: string[] = ["hello", "deno", "land"];
   const decoder = new TextDecoder();
@@ -160,30 +183,36 @@ Deno.test("toWritableCheck", async function (): Promise<void> {
   assertEquals(written, chunks);
 });
 
-Deno.test("toReadableCheck", async function (): Promise<void> {
-  const chunks: string[] = ["hello", "deno", "land"];
-  const expected = chunks.slice();
-  const readChunks: string[] = [];
+Deno.test("[io] readableStreamFromIterable() array", async function () {
+  const strings: string[] = ["hello", "deno", "land"];
   const encoder = new TextEncoder();
+  const readableStream = readableStreamFromIterable(
+    strings.map((s) => encoder.encode(s)),
+  );
 
-  // deno-lint-ignore require-await
-  async function read(p: Uint8Array): Promise<number | null> {
-    const chunk = chunks.shift();
-    if (chunk === undefined) {
-      return null;
-    } else {
-      const encoded = encoder.encode(chunk);
-      p.set(encoded);
-      return encoded.length;
-    }
-  }
-  const iter = Deno.iter({ read });
-  const writableStream = readableStreamFromAsyncIterator(iter);
-
+  const readStrings = [];
   const decoder = new TextDecoder();
-  for await (const chunk of writableStream) {
-    readChunks.push(decoder.decode(chunk));
+  for await (const chunk of readableStream) {
+    readStrings.push(decoder.decode(chunk));
   }
 
-  assertEquals(expected, readChunks);
+  assertEquals(readStrings, strings);
+});
+
+Deno.test("[io] readableStreamFromIterable() generator", async function () {
+  const strings: string[] = ["hello", "deno", "land"];
+  const readableStream = readableStreamFromIterable((function* () {
+    const encoder = new TextEncoder();
+    for (const string of strings) {
+      yield encoder.encode(string);
+    }
+  })());
+
+  const readStrings = [];
+  const decoder = new TextDecoder();
+  for await (const chunk of readableStream) {
+    readStrings.push(decoder.decode(chunk));
+  }
+
+  assertEquals(readStrings, strings);
 });