+++ path/win32.ts
@@ -13,6 +13,7 @@ import {
 import {
   _format,
   assertPath,
+  encodeWhitespace,
   isPathSeparator,
   isWindowsDeviceRoot,
   normalizeString,
@@ -979,7 +980,7 @@ export function fromFileUrl(url: string | URL): string {
  *
  *      toFileUrl("\\home\\foo"); // new URL("file:///home/foo")
  *      toFileUrl("C:\\Users\\foo"); // new URL("file:///C:/Users/foo")
- *      toFileUrl("\\\\localhost\\home\\foo"); // new URL("file://localhost/home/foo")
+ *      toFileUrl("\\\\127.0.0.1\\home\\foo"); // new URL("file://127.0.0.1/home/foo")
  * @param path to convert to file URL
  */
 export function toFileUrl(path: string): URL {
@@ -987,11 +988,11 @@ export function toFileUrl(path: string): URL {
     throw new TypeError("Must be an absolute path.");
   }
   const [, hostname, pathname] = path.match(
-    /^(?:[/\\]{2}([^/\\]+)(?=[/\\][^/\\]))?(.*)/,
+    /^(?:[/\\]{2}([^/\\]+)(?=[/\\](?:[^/\\]|$)))?(.*)/,
   )!;
   const url = new URL("file:///");
-  url.pathname = pathname.replace(/%/g, "%25");
-  if (hostname != null) {
+  url.pathname = encodeWhitespace(pathname.replace(/%/g, "%25"));
+  if (hostname != null && hostname != "localhost") {
     url.hostname = hostname;
     if (!url.hostname) {
       throw new TypeError("Invalid hostname.");